Module Stdlib.Ephemeron

include module type of struct include Stdlib.Ephemeron end

module type S = sig ... end

The output signature of the functor K1.Make and K2.Make. These hash tables are weak in the keys. If all the keys of a binding are alive the binding is kept, but if one of the keys of the binding is dead then the binding is removed.

module type SeededS = sig ... end

The output signature of the functor K1.MakeSeeded and K2.MakeSeeded.

module K1 : sig ... end

module K2 : sig ... end

module Kn : sig ... end

module GenHashTable : sig ... end